+++ aws/eks-helm/metrics-server/modules/rbac/outputs.tf
@@ -1,3 +1,0 @@
-output "agent_cluster_role_name" {
-  value = var.cluster_role_name
-}