+++ aws/eks-helm/metrics-server/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [0.1.28]() (2024-12-27)
 ### Features
 * Init `eks_rbac` sub-module to support metrics server and `var.create_eks_rbac` to enable `eks_rbac` module
-* Add `var.set_container_port`, `var.set_metrics_enabled` and `var.set_list_config` for dynamically create each set when being enabled
+* Add `var.set_container_port`, `var.set_metrics_enabled`, `var.set_rbac_create` and `var.set_list_config` for dynamically create each set when being enabled
 
 ## [0.1.21]() (2024-12-24)
 ### Features